move service and deprecate obsolete version
