+++ libs/domain/site/src/services/link/link.service.ts
@@ -1,33 +1,19 @@
-import { RouteType } from '@spryker-oryx/router';
-import { Observable } from 'rxjs';
+import {
+  LinkOptions as LinkOptionsFromRouter,
+  LinkService as LinkServiceFromRouter,
+} from '@spryker-oryx/router';
 
-export interface LinkOptions {
-  type: RouteType | string;
-  id?: string;
-  params?: Record<string, string>;
-}
+/**
+ * @deprecated since 1.4, use LinkOptions from @spryker-oryx/router instead
+ */
+export type LinkOptions = LinkOptionsFromRouter;
 
-export interface LinkService {
-  get(link: LinkOptions): Observable<string | undefined>;
+/**
+ * @deprecated since 1.4, use LinkService from @spryker-oryx/router instead
+ */
+export type LinkService = LinkServiceFromRouter;
 
-  /**
-   * Checks if the current route matches or starts with the specified URL.
-   * Asynchronously retrieves the current route and performs the check based on
-   * the exactMatch option.
-   *
-   * @param {string} url - The URL to check against the current route.
-   * @param {boolean} exactMatch - If true, checks for an exact match;
-   * otherwise, checks if the current route starts with the specified URL.
-   * @returns {Observable<boolean>} An observable that emits `true` if the current
-   * route matches or starts with the specified URL, and `false` otherwise.
-   */
-  isCurrent(url: string, exactMatch?: boolean): Observable<boolean>;
-}
-
-export const LinkService = 'oryx.LinkService';
-
-declare global {
-  interface InjectionTokensContractMap {
-    [LinkService]: LinkService;
-  }
-}
+/**
+ * @deprecated since 1.4, use LinkService from @spryker-oryx/router instead
+ */
+export const LinkService = LinkServiceFromRouter;
