+++ libs/platform/router/src/feature.ts
@@ -1,7 +1,10 @@
 import { AppFeature } from '@spryker-oryx/core';
 import { Provider } from '@spryker-oryx/di';
+import { featureVersion } from '@spryker-oryx/utilities';
 import {
+  DefaultLinkService,
   DefaultRouterService,
+  LinkService,
   routerHydration,
   RouterService,
 } from './services';
@@ -12,6 +15,9 @@ export class RouterFeature implements AppFeature {
   protected getProviders(): Provider[] {
     return [
       { provide: RouterService, useClass: DefaultRouterService },
+      ...(featureVersion >= '1.4'
+        ? [{ provide: LinkService, useClass: DefaultLinkService }]
+        : []),
       routerHydration,
     ];
   }
