+++ libs/platform/router/src/services/index.ts
@@ -1,4 +1,5 @@
 export * from './base-route';
 export * from './default-router.service';
+export * from './link';
 export * from './router-hydration';
 export * from './router.service';
