+++ libs/platform/router/src/services/link/index.ts
@@ -1,0 +1,2 @@
+export * from './default-link.service';
+export * from './link.service';
